♻️ Extract Marketing endpoint host into MarketingEndpointHost
Includes the TrendingRecomputeService background hosted service so
AllInOne picks it up automatically alongside the message endpoint.
